requirements.txt
llama-index-llms-openai
llama-index
sqlite-utils